Struct cogset::BruteScan

pub struct BruteScan<'a, P: Point + 'a> {
    // some fields omitted
}

A point collection where queries are answered via brute-force scans over the whole list.

Points are represented via their indices into the list passed to new.

Methods

impl<'a, P: Point> BruteScan<'a, P>

fn new(p: &'a [P]) -> BruteScan<'a, P>

Create a new BruteScan.